feat: clean-up the controller and extenssion
