+++ lib/hidable.dart
@@ -7,4 +7,3 @@
 library hidable;
 
 export 'src/hidable_widget.dart';
-export 'src/hidable_controller_ext.dart';
