+++ lib/src/hidable_controller.dart
@@ -7,9 +7,28 @@
 import 'package:flutter/widgets.dart';
 import 'package:flutter/material.dart';
 
+/// Simple extension to generate [HidableController] from scroll controller instance directly.
+extension HidableControllerExt on ScrollController {
+  static final hidableControllers = <int, HidableController>{};
+
+  /// Shortcut way of creating hidable controller
+  HidableController hidable(double size) {
+    // If the same instance was created before, we should keep using it.
+    if (hidableControllers.containsKey(hashCode)) {
+      return hidableControllers[hashCode]!;
+    }
+
+    return hidableControllers[hashCode] = HidableController(
+      scrollController: this,
+      size: size,
+    );
+  }
+}
+
 /// A custom wrapper for scroll controller.
 ///
-/// It have state notifiers and it's default scroll listener, Which used as default scrolling detection.
+/// Implements the main listener mehtod for [ScrollController]. 
+/// And the [sizeNotifier] for providing/updating the hideable status.
 class HidableController {
   /// The main scroll controller.
   ScrollController scrollController;
@@ -26,21 +45,17 @@ class HidableController {
 
   double li = 0.0, lastOffset = 0.0;
 
-  /// The main value notifier of widget's size.
   final sizeNotifier = ValueNotifier<double>(1.0);
-
-  /// The main value notifier of widget's stickiness.
-  final stickinessNotifier = ValueNotifier<bool>(false);
-
-  /// Adds new state value to stickiness notifier.
-  ///
-  /// Basically, used to enable/disable stickness of widget.
-  void setStickinessState(bool state) => stickinessNotifier.value = state;
 
   /// Took size factor from "li" and "size".
   double sizeFactor() => 1.0 - (li / size);
 
-  /// Default listener that detects movements on the screen, and alerts size's value-notifier.
+  /// Listener is the main "extenssion" method for the [ScrollController],
+  /// which calculates the position of scroll and decides wheter it should collapse or 
+  /// show-up the static located widget.
+  ///
+  /// The caluclation data will be emited to the [sizeNotifier]. Where 0 to 1 is the 
+  /// static-located-widget appearing status. (0 = closed) and (1 = opened).
   void listener() {
     final p = scrollController.position;
 
@@ -49,7 +64,7 @@ class HidableController {
     lastOffset = p.pixels;
 
     // If scrolled down, size-notifiers value should be zero.
-    // Can be imagined as [zero - false] | [one - true].
+    // Can be imagined as [zero = false] | [one = true].
     if (p.axisDirection == AxisDirection.down && p.extentAfter == 0.0) {
       if (sizeNotifier.value == 0.0) return;
 
@@ -72,9 +87,5 @@ class HidableController {
     sizeNotifier.value = sizeFactor();
   }
 
-  /// Closes size and stickness notifiers.
-  void close() {
-    stickinessNotifier.dispose();
-    sizeNotifier.dispose();
-  }
+  void close() => sizeNotifier.dispose();
 }
